# Conflicts:
#	GeometryThirdparty/GeometryThirdparty.csproj
#	GeometryThirdparty/PoissonReconstruction.cs
#	Pipeline/Alignment/IngestPDSImage.cs
#	Pipeline/Commands/BuildFromAlignment.cs

+++ Pipeline/TileServer/Cloud/TilingInput.cs
@@ -93,8 +93,15 @@ namespace OPS.Pipeline.TileServer
                 TilingInputChunk.Find(pipeline.DynamoContext, chunkId).Delete(pipeline, ignoreErrors, logger);
             }
 
-            pipeline.Storage(MeshUrl).DeleteObject(MeshUrl, ignoreErrors: ignoreErrors, logger: logger);
-            pipeline.Storage(ImageUrl).DeleteObject(ImageUrl, ignoreErrors: ignoreErrors, logger: logger);
+            if (!string.IsNullOrEmpty(MeshUrl))
+            {
+                pipeline.Storage(MeshUrl).DeleteObject(MeshUrl, ignoreErrors: ignoreErrors, logger: logger);
+            }
+
+            if (!string.IsNullOrEmpty(ImageUrl))
+            {
+                pipeline.Storage(ImageUrl).DeleteObject(ImageUrl, ignoreErrors: ignoreErrors, logger: logger);
+            }
 
             pipeline.DeleteDynamoItem(this, ignoreErrors, logger);
         }
