+++ Pipeline/TileServer/Cloud/TilingNode.cs
@@ -89,8 +89,16 @@ namespace OPS.Pipeline.TileServer
 
         public void Delete(PipelineCore pipeline, bool ignoreErrors = true, ILog logger = null)
         {
-            pipeline.Storage(MeshUrl).DeleteObject(MeshUrl, ignoreErrors: ignoreErrors, logger: logger);
-            pipeline.Storage(ImageUrl).DeleteObject(ImageUrl, ignoreErrors: ignoreErrors, logger: logger);
+            if (!string.IsNullOrEmpty(MeshUrl))
+            {
+                pipeline.Storage(MeshUrl).DeleteObject(MeshUrl, ignoreErrors: ignoreErrors, logger: logger);
+            }
+
+            if (!string.IsNullOrEmpty(ImageUrl))
+            {
+                pipeline.Storage(ImageUrl).DeleteObject(ImageUrl, ignoreErrors: ignoreErrors, logger: logger);
+            }
+
             pipeline.DeleteDynamoItem(this, ignoreErrors, logger);
         }
 
@@ -107,6 +115,14 @@ namespace OPS.Pipeline.TileServer
 
         public void SaveMesh(MeshImagePair pair, PipelineCore pipeline, double geometricError)
         {
+            if(pair.Image != null && !pair.Mesh.HasUVs)
+            {
+                throw new Exception("Attempting to save tiling node mesh with image but no UVs");
+            }
+            if(!pair.Mesh.HasNormals)
+            {
+                throw new Exception("Attempting to save tiling node mesh without normals");
+            }
             TemporaryFile.GetAndDelete(".ply", tmpMesh =>
             {
                 TemporaryFile.GetAndDelete(".tif", tmpImage => 
@@ -178,6 +194,22 @@ namespace OPS.Pipeline.TileServer
                         pipeline.Storage(this.ImageUrl).DownloadFile(this.ImageUrl, f);
                         img = Image.Load(f);
                     });
+                }
+                if(m == null)
+                {
+                    throw new Exception("Error loading tiling node mesh");
+                }
+                if (this.ImageUrl != null && img == null)
+                {
+                    throw new Exception("Error loading tiling node image");
+                }
+                if (img != null && !m.HasUVs)
+                {
+                    throw new Exception("Attempting to load tiling node mesh with image but no UVs");
+                }
+                if (!m.HasNormals)
+                {
+                    throw new Exception("Attempting to load tiling node mesh without normals");
                 }
                 node.AddComponent(new MeshImagePair(m, img));
                 return true;
